Re-arrange KV extensions to store stuff in the "right" place
